some more explicit casting

+++ Sampler/src/sample.c
@@ -161,8 +161,8 @@ static void sample_nopapi_omp(KernelCall *calls, size_t ncalls) {
 ////////////////////////////////////////////////////////////////////////////////
 #ifdef PAPI_ENABLED
 
-#define COUNTERS_START0 PAPI_start_counters(counters, ncounters); get_cycles(ticks0);
-#define COUNTERS_END0   get_cycles(ticks1); PAPI_stop_counters(calls[i].counters, ncounters);
+#define COUNTERS_START0 PAPI_start_counters(counters, (int) ncounters); get_cycles(ticks0);
+#define COUNTERS_END0   get_cycles(ticks1); PAPI_stop_counters(calls[i].counters, (int) ncounters);
 
 #ifndef OPENMP_ENABLED
 /** \ref sample specialization with PAPI but without OpenMP. */
